Authentication implemented

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import * as yup from 'yup';
@@ -18,10 +19,17 @@ const postSchema = yup.object({
     complete: yup.boolean().optional().default(false)
 })
 
-export async function POST(request: NextRequest) { 
+export async function POST(request: NextRequest) {
+    
+    const user = await getUserSessionServer();
+
+    if( !user ) {
+        return NextResponse.json('Unauthorized', {status: 401});
+    }
+    
     try {
         const { description, complete } = await postSchema.validate(await request.json());
-        const todo = await prisma.todo.create({data: { description, complete }})
+        const todo = await prisma.todo.create({data: { description, complete, userId: user.id }})
         return NextResponse.json(todo);
     } catch (error) {
         return NextResponse.json(error, { status: 500 });
@@ -29,10 +37,17 @@ export async function POST(request: NextRequest) {
 }
 
 export async function DELETE(request: NextRequest) { 
+    const user = await getUserSessionServer();
+
+    if( !user ) {
+        return NextResponse.json('Unauthorized', {status: 401});
+    }
+
     try {
         const deletedTodos = await prisma.todo.deleteMany({
             where: {
-                complete: true
+                complete: true,
+                userId: user.id
             }
         })
         return NextResponse.json(deletedTodos);
